requirements.txt
flask
torch
torchvision
pillow
timm
huggingface_hub